testapp/requirements.txt
-e ..[slack]
pytest==6.2.5
pytest-cov==2.12.1
pytest-django==4.4.0
black==20.8b1
